Prefer wallet methods, usage flags when connecting to wallet.
This prevents flag mismatch errors when attempting to invoke a method
that is implemented by both dcrd and dcrwallet but with different set
of request and/or return params. Also ensures the appropriate help msg
is displayed, for methods with different help msgs defined in dcrd and
dcrwallet.

If the `--wallet` flag is passed, method usage flags defined by wallet
are used (falling back to dcrd usage flags, if the method is not impl.
by wallet).

The previous behaviour of using dcrd method flags first (with wallet
fallback for dcrd-unimplemented methods) is retained when user does
not explicitly set the wallet flag to true.